Add UART RX receiver module to the project
This module can receive information through the serial port RX. There is a possibility of flexide setting using the parameters in the module header. The module sends in parallel code each received data packet and set the confirmation pulse on a separate bus. In addition, the module can recognize a packet transmission error for several reasons. Firstly, an error occurs when the parity bit is incorrect. Secondly, the error occurs when the stop bit is incorrectly received.